Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "jwt>=1.4.0",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "httpx>=0.28.1",
]
